Commit 3
Fixed a issue with the persistence and some issues with the ingredient entity file. Fixed some comments. Figured out that RecipeHelper's SQL statments work incorrect and fixed them. Finished the add-recipe.jsp and the view-all-recipes.jsp. Moved web pages to the correct folder. Next is working on making the edit and delete work properly.