Commit 5
Added CSS to each of the webpages to make the website more appealing. Added a welcome message and explanation for the user in the index. Fixed a mistake in the RecipeHelper. Added comments for partner. Fixed previous comments throughout all files. Updated the nvigationServlet slightly. Fixed an issue with the editRecipeServlet (Forgot to remove  some functionalilty that I didn't end up using) caused the date to always be set to the current date. From here I will have my partner finish what's left of the project, anymore commits from me will be assistance with her part, or added functionalilty if there is time left.